List surveys
This file lists the surveys the Census Bureau has for the user to choose from. Some ways to expand (in separate branches) are:

1. Web scrape for new survey names at https://www.census.gov/programs-surveys/surveys-programs.html when the program runs.
2. Add a description to each survey when given user input. 

First, though, we will hardcode some of their biggest surveys in to the list (because they will likely stay in the Census offerings) and expand from there.